feat: restructure for pynvim 0.4.3 backwards compatibility
* move files under subdirectory

* change to use local relative imports

* fix healthcheck to support pynvim 0.4.3 and use correct python version

* Fix imports again
