+++ lua/CopilotChat/health.lua
@@ -53,7 +53,7 @@ function M.check()
     return
   end
 
-  file:write('import pynvim; print(pynvim.__version__)')
+  file:write('import pynvim; v = pynvim.VERSION; print("{0}.{1}.{2}".format(v.major, v.minor, v.patch))')
   file:close()
 
   -- Run the temporary Python script and capture the output
@@ -67,7 +67,7 @@ function M.check()
   -- Delete the temporary Python script
   os.remove(temp_file)
 
-  if pynvim_version ~= '0.5.0' then
+  if vim.version.lt(pynvim_version, "0.4.3") then
     warn('pynvim version ' .. pynvim_version .. ' is not supported')
   else
     ok('pynvim version ' .. pynvim_version .. ' is supported')
