+++ rplugin/python3/CopilotChat/__init__.py
@@ -1,0 +1,1 @@
+from .copilot_plugin import CopilotPlugin as CopilotPlugin
