+++ rplugin/python3/CopilotChat/copilot_plugin.py
@@ -1,7 +1,7 @@
 import pynvim
-from handlers.inplace_chat_handler import InPlaceChatHandler
-from handlers.vsplit_chat_handler import VSplitChatHandler
-from mypynvim.core.nvim import MyNvim
+from CopilotChat.handlers.inplace_chat_handler import InPlaceChatHandler
+from CopilotChat.handlers.vsplit_chat_handler import VSplitChatHandler
+from CopilotChat.mypynvim.core.nvim import MyNvim
 
 PLUGIN_MAPPING_CMD = "CopilotChatMapping"
 PLUGIN_AUTOCMD_CMD = "CopilotChatAutocmd"
