+++ rplugin/python3/CopilotChat/handlers/vsplit_chat_handler.py
@@ -1,6 +1,6 @@
-from handlers.chat_handler import ChatHandler
-from mypynvim.core.buffer import MyBuffer
-from mypynvim.core.nvim import MyNvim
+from CopilotChat.handlers.chat_handler import ChatHandler
+from CopilotChat.mypynvim.core.buffer import MyBuffer
+from CopilotChat.mypynvim.core.nvim import MyNvim
 
 
 class VSplitChatHandler(ChatHandler):
